// frontend/src/expense/store/types.ts
export enum Types {
    GET_EXPENSES_SUCCESS = 'GET_EXPENSES_SUCCESS',
    GET_EXPENSES_ERROR = 'GET_EXPENSES_ERROR',
    CREATE_EXPENSES_SUCCESS = 'CREATE_EXPENSES_SUCCESS',
    CREATE_EXPENSES_ERROR = 'CREATE_EXPENSES_ERROR',
    UPDATE_EXPENSES_SUCCESS = 'UPDATE_EXPENSES_SUCCESS',
    UPDATE_EXPENSES_ERROR = 'UPDATE_EXPENSES_ERROR',
    DELETE_EXPENSES_SUCCESS = 'DELETE_EXPENSES_SUCCESS',
    DELETE_EXPENSES_ERROR = 'DELETE_EXPENSES_ERROR',
    SIGN_IN_PENDING = 'SIGN_IN_PENDING',
    SIGN_IN_SUCCESS = 'SIGN_IN_SUCCESS',
    SIGN_IN_ERROR = 'SIGN_IN_ERROR',
    SIGN_UP_PENDING = 'SIGN_UP_PENDING',
    SIGN_UP_SUCCESS = 'SIGN_UP_SUCCESS',
    SIGN_UP_ERROR = 'SIGN_UP_ERROR'
}